Rename modify to manage
