+++ README.md
@@ -35,15 +35,15 @@ cargo install scoob --locked
 First, you'll want to create a secrets file:
 
 ```bash
-scoob modify --create ./secrets/dev.yml
+scoob manage ./secrets/dev.yml
 ```
 
 This will open your editor with an example Scoob configuration file. A pair of public and secret keys will auto-generated and provided in the file. Make sure you don't commit these into your repository, and instead replace them with values provided dynamically via environment variables. When you close your editor, all of the values under `configuration:` will be encrypted, and the file will be written to disk.
 
-At a later point, you can add additional secrets by running the following command:
+At a later point, you can add additional secrets by running the same command:
 
 ```bash
-scoob modify --edit ./secrets/dev.yml
+scoob manage ./secrets/dev.yml
 ```
 
 We recommend creating a separate secrets file for development and production. This way, you can keep your production keys separate.
