+++ src/main.rs
@@ -1,56 +1,26 @@
 mod config;
 mod encryption;
-mod modify;
+mod manage;
 mod start;
 
 use crate::config::*;
 use crate::encryption::*;
-use crate::modify::*;
+use crate::manage::*;
 use crate::start::*;
 use colored::Colorize;
 use std::alloc::System;
-use std::path::PathBuf;
 use structopt::StructOpt;
 
 #[global_allocator]
 static A: System = System;
 
-#[derive(Debug, PartialEq, StructOpt)]
-pub enum SubCommand {
-    #[structopt(external_subcommand)]
-    Other(Vec<String>),
-}
-
-#[derive(Debug, StructOpt)]
-pub struct Modify {
-    /// Edit the provided configuration file
-    #[structopt(short, long)]
-    edit: bool,
-    /// Create the configuration file
-    #[structopt(short, long)]
-    create: bool,
-    /// Path to the scoob configuration file
-    #[structopt(parse(from_os_str))]
-    file: PathBuf,
-}
-
-#[derive(Debug, StructOpt)]
-pub struct Start {
-    /// Path to the scoob configuration file
-    #[structopt(parse(from_os_str))]
-    file: PathBuf,
-    /// The sub-command that you wish to run
-    #[structopt(subcommand)]
-    sub_command: SubCommand,
-}
-
 #[derive(Debug, StructOpt)]
 #[structopt(name = "scoob", about = "A secrets management tool.")]
 enum Opt {
-    /// Open a scoob configuration file for modification
-    Modify(Modify),
+    /// Manage a scoob configuration file
+    Manage(Manage),
 
-    /// Runs a command after loading scoob configuration into the environment
+    /// Runs a command after loading scoob secrets into the environment
     Start(Start),
 }
 
@@ -62,9 +32,9 @@ fn main() {
     let cli = Opt::from_args();
 
     let result = match &cli {
-        Opt::Modify(c) => modify(&c),
+        Opt::Manage(c) => manage(&c),
         Opt::Start(c) => {
-            let start_result = start(c);
+            let start_result = start(&c);
 
             match start_result {
                 Ok(status) => std::process::exit(status),
